ims-master/v3/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-createsuperuserwithpassword==2.0.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-stdimage==5.0.2
gunicorn==20.0.4
Pillow==7.0.0
progressbar2==3.47.0
psycopg2-binary==2.8.4
python-utils==2.3.0
pytz==2019.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
